[Fix] 星を半分で表示できるように変更

+++ db/migrate/20220702134705_remove_rate_from_books.rb
@@ -1,5 +1,5 @@
 class RemoveRateFromBooks < ActiveRecord::Migration[6.1]
   def change
-    remove_column :books, :rate, :float
+    remove_column :books, :rate, :integer
   end
 end
