+++ db/migrate/20220702134800_add_rate_to_books.rb
@@ -1,5 +1,5 @@
 class AddRateToBooks < ActiveRecord::Migration[6.1]
   def change
-    add_column :books, :rate, :integer
+    add_column :books, :rate, :float
   end
 end
